Soft delete funcionando

+++ app/Http/Controllers/UserController.php
@@ -88,12 +88,18 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
-        $user->delete();
+        if (auth()->user()->id != $user->id)
+            $user->delete();
         return redirect()->route('user.index');
     }
 
-    public function revive(User $user)
+    public function revive($user_id)
     {
+        $user_to_revive = (object)User::withTrashed()->find($user_id)->getAttributes();
+        $user = new User();
+        foreach ($user_to_revive as $key => $value) {
+            $user->$key = $value;
+        }
         $user->restore();
         return redirect()->route('user.index');
     }
